MAINT: Update router for landing page
