+++ src/pages/home.tsx
@@ -1,42 +1,92 @@
 import * as React from 'react'
-import Image from 'next/image'
-import Link from 'next/link'
+import Box from '@mui/material/Box'
 import Button from '@mui/material/Button'
-import Grid from '@mui/material/Grid'
-import Stack from '@mui/material/Stack'
+import Container from '@mui/material/Container'
+import Fab from '@mui/material/Fab'
 import Typography from '@mui/material/Typography'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { faAdd } from '@fortawesome/free-solid-svg-icons'
+import dayjs from 'dayjs'
+import { useRouter } from 'next/router'
 
-type Props = {
-  image: string // The path of image
-}
-const Home: React.FC<Props> = ({ image }) => {
+import PlansList from 'components/modules/PlansList'
+import { PlanDB } from 'contexts/CurrentPlanProvider'
+import { useAuthentication } from 'hooks/firebase/useAuthentication'
+import { usePlans } from 'hooks/usePlan'
+import { useConfirm } from 'hooks/useConfirm'
+import PlanningLayout from 'components/layouts/PlaningLayout'
+
+const UserHome = () => {
+  const router = useRouter()
+  const [user] = useAuthentication()
+  const actions = usePlans()
+  const confirm = useConfirm()
+  const [plans, setPlans] = React.useState<Array<PlanDB>>([])
+  const [nextPlan, setNextPlan] = React.useState<PlanDB | null>(null)
+
+  React.useEffect(() => {
+    if (user) {
+      actions.fetch().then((results) => {
+        setPlans(results || [])
+      })
+    } else {
+      setPlans([])
+    }
+  }, [actions, user])
+
+  React.useEffect(() => {
+    // 将来の旅行計画の中から、最も近い旅行を表示する
+    const today = new Date()
+    const sortedFeaturesDesc = plans
+      .filter(({ data }) => data.end > today)
+      .sort((a, b) => dayjs(b.data.start).diff(a.data.start))
+    setNextPlan(sortedFeaturesDesc.shift() || null)
+  }, [plans])
+
+  const handleClick = async () => {
+    if (!user) {
+      try {
+        await confirm({
+          title: 'Create Plan?',
+          description:
+            'NOTE: You are a guest user, you cannot save plan. Please create account or login if you want to save.',
+          allowClose: true,
+        })
+      } catch {}
+    }
+    router.push('new')
+  }
+
   return (
-    <Grid container direction={['column-reverse', 'column-reverse', 'row']}>
-      <Grid item md={4} xs={12}>
-        <Stack alignItems="center" spacing={4} sx={{ mx: 4 }}>
-          <Typography variant="h3">キャッチコピー</Typography>
-          <Typography variant="h5">
-            Short description, Short description, Short description, Short
-            description, Short description, Short description, Short
-            description, Short description, Short description, Short
-            description, Short description, Short description,
-          </Typography>
-          <Link href="/new" passHref>
-            <Button variant="contained">プランを作成する</Button>
-          </Link>
-        </Stack>
-      </Grid>
-      <Grid item md={8} xs={12}>
-        <Image src={image} width="100%" height={100} alt="hero image" />
-      </Grid>
-    </Grid>
+    <PlanningLayout>
+      {nextPlan && (
+        <Box style={{ height: '20%', backgroundColor: '#aaaaaa50' }}>
+          <Typography>Next trip</Typography>
+        </Box>
+      )}
+      <Container
+        maxWidth="lg"
+        sx={{ my: 4, display: 'flex', justifyContent: 'center' }}>
+        {plans.length > 0 ? (
+          <PlansList plans={plans} />
+        ) : (
+          <Button variant="contained" onClick={handleClick}>
+            Plan Your Travel
+          </Button>
+        )}
+      </Container>
+      <Fab
+        onClick={handleClick}
+        color="primary"
+        sx={{
+          position: 'fixed',
+          right: 16,
+          bottom: 16,
+        }}>
+        <FontAwesomeIcon icon={faAdd} size="lg" />
+      </Fab>
+    </PlanningLayout>
   )
 }
 
-export const getStaticProps = () => {
-  return {
-    props: { image: '/' },
-  }
-}
-
-export default Home
+export default UserHome
